+++ tencentcloud/vpc/v20170312/models.go
@@ -11320,10 +11320,10 @@ func (r *HaVipDisassociateAddressIpResponse) FromJsonString(s string) error {
 
 type IKEOptionsSpecification struct {
 
-	// 加密算法，可选值：'3DES-CBC', 'AES-CBC-128', 'AES-CBS-192', 'AES-CBC-256', 'DES-CBC'，默认为3DES-CBC
+	// 加密算法，可选值：'3DES-CBC', 'AES-CBC-128', 'AES-CBS-192', 'AES-CBC-256', 'DES-CBC'，'SM4', 默认为3DES-CBC
 	PropoEncryAlgorithm *string `json:"PropoEncryAlgorithm,omitempty" name:"PropoEncryAlgorithm"`
 
-	// 认证算法：可选值：'MD5', 'SHA1'，默认为MD5
+	// 认证算法：可选值：'MD5', 'SHA1'，'SHA-256' 默认为MD5
 	PropoAuthenAlgorithm *string `json:"PropoAuthenAlgorithm,omitempty" name:"PropoAuthenAlgorithm"`
 
 	// 协商模式：可选值：'AGGRESSIVE', 'MAIN'，默认为MAIN